Update Payment || Request || Transkasi

+++ database/migrations/2022_08_05_105404_create_promos_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->string('category_id',10)->nullable();
             $table->integer('max_user')->nullable();
             $table->integer('use_user')->nullable()->default('0');
-            $table->bigInteger('promo_persen')->nullable();
+            $table->integer('promo_persen')->nullable();
             $table->bigInteger('promo_nominal')->nullable();
             $table->date('tgl_mulai');
             $table->date('tgl_kadaluarsa');
